tidy/export helper interfaces (#88)
