+++ src/interfaces/Statement/MultiPart.ts
@@ -1,0 +1,4 @@
+export interface MultiPart {
+  header: { "Content-Type": string };
+  blob: Blob;
+}
