+++ src/interfaces/Statement/Part.ts
@@ -1,0 +1,3 @@
+import { Statement } from ".";
+
+export type Part = Statement | unknown;
