+++ src/interfaces/XAPI/XAPILaunchData.ts
@@ -1,0 +1,6 @@
+import { Actor } from "../Statement";
+
+export interface XAPILaunchData {
+  endpoint: string;
+  actor: Actor;
+}
