+++ src/interfaces/XAPI/XAPILaunchParameters.ts
@@ -1,0 +1,5 @@
+export interface XAPILaunchParameters {
+  xAPILaunchKey?: string;
+  xAPILaunchService?: string;
+  encrypted?: string;
+}
