+++ src/interfaces/XAPI/index.ts
@@ -2,3 +2,6 @@ export * from "./Query";
 export * from "./RequestParams";
 export * from "./Resource";
 export * from "./StatementsResponse";
+export * from "./TinCanLaunchData";
+export * from "./XAPILaunchData";
+export * from "./XAPILaunchParameters";
